mgr/dashboard/telemetry: reduce telemetry dashboard preview size
Keeping perf channel metrics on the telemetry preview form may cause
laggy loading times and slow behavior in the text box if the report
is particularly long. To solve this, we can remove performance metrics
from the preview form and notify the users that they may` still download
a full report upon selecting the "Download" button.

Fixes: https://tracker.ceph.com/issues/53831
Signed-off-by: Laura Flores <lflores@redhat.com>