doc/mgr: name data source in "Man Install & Config"
This PR specifies that the data source must be set to
be "Dashboard1" when you configure Grafana and Prometheus
manually.

This is cherry-picked from this PR:

https://github.com/ceph/ceph/pull/44564/

PR#44564 got caught in a broken dashboard test.

This is a fixup of another PR which was created by Dr
Jake Grimmett. This is that PR:

https://github.com/ceph/ceph/pull/44150/

Credit goes to Dr Jake Grimmett of Cambridge.

Signed-off-by: Zac Dover <zac.dover@gmail.com>
(cherry picked from commit 987713da331e941c05d114f1e167c21d1e7896eb)