librbd: diff-iterate reports incorrect offsets if whole_object=true
It turns out that in octopus both fast-diff and list-snaps (slow)
modes were broken.  As long as whole_object=true, the same incorrect
offset was reported in both modes.  The fast-diff mode is fixed in
in previous commit.

This is an octopus-only patch for list-snaps mode.  In pacific this
issue was addressed with 4429ed4f3f4c ("librbd: switch diff iterate
API to use new snaps list dispatch methods").

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>