mgr/dashboard: BDD approach for the dashboard cephadm e2e
Files under the directory cypress/integration/common/* will contain
common specs which can be used on all the .feature files. We can change
the common directory to cypress/integration/* from the package.json, but
if we do that now then we'll need to take care of all the absolute
import in that path. So for now at least that's not a good choice.

The bug in the cypress-browserify-preprocessor which doesn't allow to
take our tsconfig.json file forced me to go with relative imports rather
than the absolute import. We'll need to wait for this to be fixed before
 changing all our tests to BDD.

Fixes: https://tracker.ceph.com/issues/54224
Signed-off-by: Nizamudeen A <nia@redhat.com>
(cherry picked from commit 79f8545deba48a2bac6816caa7281e2198f8e5b8)