kill: get machine type from paddles for the run
When calling kill for a run which is still in beanstalkd
queue there is no job directories created in archive
and there is no way to find out which machine type to use
the teuthology-kill reports that you must manually provide
machine type for the run, which is often borring and quite
inconvenient.

Instead of torturing user to recall a tube for the run name
we just ask paddles if it has anything logged and use the
machine type from what we receive from it.

Signed-off-by: Kyr Shatskyy <kyrylo.shatskyy@suse.com>