rgw_sal_motr: [EOS-27563] Initiate default values for Zonegroup (#78)
* rgw_sal_motr: [EOS-27563] Initiate default values for zonegroup

Initialized the ZoneGroup id and api_name with default values.
Now on while creating the bucket the default zonegroup_id is assigned to the bucket, and
on requesting location of the bucket via cmds like `s3cmd info` or `s3api get-bucket-location`,
the default zonegroup api_name is returned.

Signed-off-by: Sumedh A. Kulkarni <sumedh.a.kulkarni@seagate.com>

* rgw_sal_motr: [EOS-27563] add a TODO to fetch values from config

The Default Zonegroup Params should be fetched from provisioner config.
Added a TODO for same.

Signed-off-by: Sumedh A. Kulkarni <sumedh.a.kulkarni@seagate.com>