mgr/progress: 0 affected PG doesn't break test case
Currently in qa/tasks/mgr/test_progress.py
we don't have a case where PG recovery event
doesn't get triggered when osd is marked in/out
due to the osd not being in any of the PGs acting
sets in both old osd map and new osd map.
This causes rados:mgr:task/progress.yaml to fail
at random sometimes.

Therefore, in the case where 0 PGs are affected
when making osd(s) in/out, we genereate this
as an event and immediately add this to the completed
events of the progress module. This allows the
test cases in qa/tasks/mgr/test_progress.py
to check for 0 PGs are affected event, and
evaluate the correctness accurately.

This commit adds to the unittest in:
src/pybind/mgr/progress/test_progress.py.
We add a test case to the function,
_osd_in_out() to evaluate whether a
"0 PGs affected by osd.x being marked in/out"
gets created in completed events.

Signed-off-by: Kamoltat <ksirivad@redhat.com>