qa/tasks/mgr/test_progress.py: test for 0 affected PG
Removed _osd_in_out_events_count() entirely and
modified _get_osd_in_out_events() to take in
more arguments that would enable the caller
ask for both 'in progress' and 'completed'
events.

Lastly, we added _check_no_pgs_affected()
to catch the case where there were no PGs
being affected by marking osd in/out, hence
we simply skip the process of waiting for
an event to pop up and continue on with
the rest of the tests. This has been
added to any tests that involves marking osd(s)
in/out and waitng for events to pop up.

Signed-off-by: Kamoltat <ksirivad@redhat.com>