doc/start: include A. D'Atri's hardware-recs recs
This PR restores material about partition alignment
and material about separating OS and OSD data that
was removed in an earlier rewrite. The restoration
of this information was requested by Anthony D'Atri in
https://github.com/ceph/ceph/pull/45123/

This PR also includes several refinements to the language
that could not be made to this text until now, owing to my
(Zac's) ignorance and illiteracy.

I call upon Mark Nelson (and anyone else with sufficient
command of the current state of storage technology) to advise
me on whether the Ceph Foundation feels comfortable in the year
2022 referring to QLC as an emerging technology.

Signed-off-by: Zac Dover <zac.dover@gmail.com>

(squash) more notes and revisions

Signed-off-by: Zac Dover <zac.dover@gmail.com>
One conflict had to be resolved. It contained the confval
references that were added to the documentation yesterday
in PR#45139.

(cherry picked from commit 07a1f7fe0463c5ef4424646e811626ff4bbf729c)

Signed-off-by: Zac Dover <zac.dover@gmail.com>

(squash) remove confvals

The :confval: directives that were in this
file cannot be parsed in the Pacific branch.
They can therefore not be backported to the
Pacific branch and hence must be removed.

This commit removes those directives.

It will also be squashed, so if you are
reading this, you are digging deep for
some reason.

Signed-off-by: Zac Dover <zac.dover@gmail.com>