mds: add long snap name flag in LeaseStat for fscrypt
For the long snapshot names, which are from the parent CInodes,
the format will be "_${ENCRYPTED-NAME}_${PARENT-INO}". In kclient
side we need to base64 decode it to human readable format of
"_${DENCRYPTED-NAME}_${PARENT-INO}" if the encrypt key exists. So
we need to know whether the current snapshot name is a long snap
name.

Signed-off-by: Xiubo Li <xiubli@redhat.com>