os/bluestore: Give fsck/repair ability to delete broken objects
Each object's metadata is stored in RocksDB as 1 or more keys.
If some keys are missing there fsck/repair used to leave it alone.
If repair finds an object that is missing any key then it deletes
all keys that belonged to this object. It relies on other parts of
fsck/repair to clean allocations/shared blobs/omaps.

Signed-off-by: Adam Kupczyk <akupczyk@redhat.com>