Add skipif_teuthology_process
There are a couple unit tests that seem to be fundamentally unable to
be used properly if they are run as part of a teuthology process, as
opposed to directly inside py.test. This fixture makes it easy to skip
those tests in such an environment without skipping them elsewhere.

Signed-off-by: Zack Cerza <zack@redhat.com>