rgw/rgw_common: Keyname should not include version_id in case of                 unversioned bucket.
Problem : If version id parameter is passed as null to delete-object
API in case of unversioned bucket for deletion of simple object,
then object deletion does not happen. This case is valid only when
Motr is used as backend for rgw.

Implementation : In case of unversioned bucket the to_str() function
of rgw_obj_key returns key in the following format <obj_key>[null].
To avoid this a new function get_key_name_string() is added which
returns key in the format <obj_key> if instance variable of
rgw_obj_key is set to "null" string.

Other half of this PR: https://github.com/Seagate/cortx-rgw/pull/114

Signed-off-by: Deepak Mahale <deepak.mahale@seagate.com>