librbd: readv/writev fix iovecs length computation overflow
iovec have unsigned length (size_t) and we compute total length
of all iovecs by adding them to a signed length variable (ssize_t).
While the old code checked if the resulting length was negative on
overflow, the case where length is positive after overflow was not
checked.
This patch fixes the total length computation by changing ssize_t to
size_t and checking for overflows with __builtin_add_overflow.
Io::aio_write and Io::aio_read take uint64_t for length, so no need
to use signed for computation.

Additionally, this patch fixes the case where some iovecs have been
added to the bufferlist and the aio completion has been blocked, but
adding an additional iovec fails because of overflow. This leads to
the UserBufferDeleter trying to unblock the completion on destruction
of the bufferlist but asserting because the completion was never
armed. We avoid this by first computing the total length and checking
for overflows and iovcnt before adding them to the bufferlist.

Signed-off-by: Jonas Pfefferle <pepperjo@japf.ch>