doc/start: remove osd stub from hardware recs
This PR removes a stub section called "OSDS (CEPH-OSD)"
from the hardware-recommendations.rst page. This section
had no content in it.

This PR is made in lieu of a backport. (I had thought
that this section was removed by another PR that was
a backport, but I was wrong. So this fixes it.)

Signed-off-by: Zac Dover <zac.dover@gmail.com>