mds,client: validate pool against pool ids as well as pool names
Problem:
If the pool is a numeric string (eg. 23134), the current validation assumes
that it is a pool id and looks up the list of pool ids and fails validation
if the pool is not found amongst the list of pool ids.

Solution:
Since a string of digits can also be used as a pool name, the pool is validated
against list of pool names as well if it doesn't match any of the pool ids.

Fixes: https://tracker.ceph.com/issues/55165

Signed-off-by: wangxinyu <wangxinyu@inspur.com>