build: LTO can cause false positives in cmake tests
In particular the cmake/modules/CheckCxxAtomic.cmake test will
report success for HAVE_CXX11_ATOMIC on s390x in Fedora.

Note that the Fedora package builds use the %cmake, %cmake_build, and
%cmake_install macros defined in /usr/lib/rpm/macros.d/macros.cmake
which add f -flto=auto -ffat-lto-objects to the CFLAGS and CXXFLAGS.

The HAVE_CXX11_ATOMIC test incorrectly reports success when LTO
determines that atomic16() is unused and optimizes it out, and it
successfully links, even though one might naively think it shouldn't.
Thus the test doesn't perform the subsequent test for HAVE_LIBATOMIC.
The build then proceeds but fails later on when attempts to link
certain programs fail with undefined references to __atomic_load_16,
__atomic_store_16, and __atomic_compare_exchange_16'

https://tracker.ceph.com/issues/54514

Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>