rgw: generation numbers in data sync are not optional<>
data sync was passing 'empty' generation numbers into
`RGWDataSyncSingleEntryCR` when triggered by full sync. `RGWSyncBucketCR`
treated these empty generation numbers as a request to sync the
current generation tracked in `bucket_status.incremental_gen`

`RGWDataSyncSingleEntryCR` uses the 'bucket sync cache' to prevent itself
from spawning concurrent syncs on the same entry. so when a new request
conflicts with an existing one, it must decide which one to throw away.
the use of an optional generation number made it impossible for
`RGWDataSyncSingleEntryCR` to enforce a strict ordering between these
obligations, so it was just as likely to throw away the wrong one

replace the 'empty' generation numbers from data full sync with the
value 0. instead of requesting sync of the 'current' generation, this
obligation with generation 0 is potentially much less restrictive:

if the bucket has not ever been resharded, its only log generation would
be 0, and this change would have no effect: the bucket's current would
match both the 'empty' generation and an explicit generation 0

but if the bucket has been resharded, an obligation of `gen=0` only
requires that we've made it past 'bucket full sync'. once we finish full
sync and transition to incremental sync on its latest generation, we'll
do the comparison with the input `gen=0` and return success after
logging "requested sync of past generation 0 < Y, returning success".
after 'data full sync' switches to incremental, we'll eventually see the
datalog entries for any changes to later generations and trigger their
sync then

Signed-off-by: Casey Bodley <cbodley@redhat.com>