mds: make all the subdirs inherit the encyrption info from their parent
Since in a hierarchy only the oldest ancester allowed to set the
encryption info,g and all the subdirs in the same hierarchy will
inherit the encryption info from it, and this encryption info
will be used to encyrpt the dentry names only.

This could allow use to support the snapshot name more easy when
handling the long snap names.

And for files in the hierarchy we will allow it	to have its own
dedicated encryption info, which will be used to encrypt the file
contents.

And for link files, we don't care.

Signed-off-by: Xiubo Li <xiubli@redhat.com>